+++ migrations/20210504191052-create-assenza.js
@@ -6,10 +6,10 @@ module.exports = {
                 allowNull: false,
                 autoIncrement: true,
                 primaryKey: true,
-                type: Sequelize.INTEGER
+                type: Sequelize.BIGINT
             },
             userId: {
-                type: Sequelize.BIGINT(100)
+                type: Sequelize.BIGINT
             },
             data: {
                 type: Sequelize.DATE
